+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMW99M9KTZZBB21SSZ03MBCG"
+        "01KMYVYYVTXVPVMJ94V63V2BEV"
     )
     r"""The unique identifier of the data source"""