+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMW99M9JWH5TXGQACE6PTM85"
+        "01KMYVYYVTVR3BPH98C97S3QFZ"
     )
     r"""The unique identifier of the data source"""