+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-29T07:52:33.432Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-30T07:57:15.372Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")