+++ packages/orq-rc/RELEASES.md
@@ -6669,3 +6669,13 @@ Based on:
 - [python v4.7.0-rc.26] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.26] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.26 - packages/orq-rc
+
+## 2026-03-30 08:00:17
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.3 (2.869.25) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.27] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.27] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.27 - packages/orq-rc